INT-3297: Added `Loader.withVersion` for using miniature to preload TinyMCE
Also removed `deleteTinyMCE` from Loader.tsx and moved it to
LoadTinyTest.ts
