+++ src/test/ts/alien/Loader.tsx
@@ -1,10 +1,11 @@
-import { Arr, Fun, Global, Optional, Strings } from '@ephox/katamari';
-import { Attribute, Remove, SelectorFilter, SugarElement, SugarNode } from '@ephox/sugar';
+import { Fun, Optional } from '@ephox/katamari';
+import { SugarElement, SugarNode } from '@ephox/sugar';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { Editor, IAllProps, IProps } from '../../../main/ts/components/Editor';
+import { Editor, IAllProps, IProps, Version } from '../../../main/ts/components/Editor';
 import { Editor as TinyMCEEditor } from 'tinymce';
-import { ScriptLoader } from '../../../main/ts/ScriptLoader2';
+import { before, context } from '@ephox/bedrock-client';
+import { VersionLoader } from '@tinymce/miniature';
 
 // @ts-expect-error Remove when dispose polyfill is not needed
 Symbol.dispose ??= Symbol('Symbol.dispose');
@@ -28,28 +29,12 @@ export interface ReactEditorContext extends Context, Disposable {
   remove(): void;
 }
 
-const deleteTinymce = () => {
-  ScriptLoader.reinitialize();
-
-  delete Global.tinymce;
-  delete Global.tinyMCE;
-  const hasTinymceUri = (attrName: string) => (elm: SugarElement<Element>) =>
-    Attribute.getOpt(elm, attrName).exists((src) => Strings.contains(src, 'tinymce'));
-
-  const elements = Arr.flatten([
-    Arr.filter(SelectorFilter.all('script'), hasTinymceUri('src')),
-    Arr.filter(SelectorFilter.all('link'), hasTinymceUri('href')),
-  ]);
-
-  Arr.each(elements, Remove.remove);
-};
-
 export const render = async (props: Partial<IAllProps> = {}, container: HTMLElement = getRoot()): Promise<ReactEditorContext> => {
   const originalInit = props.init || {};
   const originalSetup = originalInit.setup || Fun.noop;
   const ref = React.createRef<Editor>();
 
-  const context = await new Promise<Context>((resolve, reject) => {
+  const ctx = await new Promise<Context>((resolve, reject) => {
     const init: IProps['init'] = {
       ...originalInit,
       setup: (editor) => {
@@ -87,16 +72,30 @@ export const render = async (props: Partial<IAllProps> = {}, container: HTMLElem
 
   const remove = () => {
     ReactDOM.unmountComponentAtNode(container);
-    deleteTinymce();
   };
 
   return {
-    ...context,
+    ...ctx,
     /** By rendering the Editor into the same root, React will perform a diff and update. */
     reRender: (newProps: IAllProps) => new Promise<void>((resolve) =>
-      ReactDOM.render(<div><Editor apiKey='no-api-key' ref={context.ref} {...newProps} /></div>, container, resolve)
+      ReactDOM.render(<div><Editor apiKey='no-api-key' ref={ctx.ref} {...newProps} /></div>, container, resolve)
     ),
     remove,
     [Symbol.dispose]: remove
   };
+};
+
+type RenderWithVersion = (
+  props: Omit<IAllProps, 'cloudChannel' | 'licenseKey' | 'tinymceScriptSrc'>,
+  container?: HTMLElement | HTMLDivElement
+) => Promise<ReactEditorContext>;
+
+export const withVersion = (version: Version, fn: (render: RenderWithVersion) => void): void => {
+  context(`TinyMCE (${version})`, () => {
+    before(async () => {
+      await VersionLoader.pLoadVersion(version);
+    });
+
+    fn(render as RenderWithVersion);
+  });
 };
